Added uniform bw (16-bit) support for posits

+++ tools/SeeDot/seedot/Predictor/predictors.h
@@ -4,10 +4,10 @@
 #pragma once
 
 void seedotFixed(MYINT** X, int32_t* res);
-void seedotPosit(float** X, int32_t* res);
+void seedotPosit(float** X, float* res);
 void seedotFloat(float** X, float* res);
 void seedotFixedSwitch(int i, MYINT** X, int32_t* res);
-void seedotPositSwitch(int i, float** X, int32_t* res);
+void seedotPositSwitch(int i, float** X, float* res);
 
 extern const int switches;
 extern const int positSwitches;
